+++ boilerplate/templatetags/boilerplate.py
@@ -54,7 +54,7 @@ def form_app_name(value):
   """
   Return the app name of a form model
   """
-  return value._meta.model._meta.verbose_name
+  return value._meta.model._meta.app_config.verbose_name.title()
 
 @register.filter
 def form_app_url(value):
@@ -126,7 +126,7 @@ def model_app_name(value):
   """
   Return the app verbose name of an object
   """
-  return value._meta.verbose_name
+  return value._meta.app_config.verbose_name.title()
 
 @register.filter
 def model_app_url(value):
@@ -152,7 +152,7 @@ def queryset_app_name(value):
   """
   Return the app verbose name of a queryset
   """
-  return value.model._meta.verbose_name
+  return value.model._meta.app_config.verbose_name.title()
 
 @register.filter
 def queryset_app_url(value):