+++ HISTORY.rst
@@ -3,6 +3,13 @@
 History
 -------
 
+0.1.3 (2016-06-12)
+++++++++++++++++++
+* Fix: CRUD Messages Mixin conflict with Extra Forms and Formsets Mixin
+* Fix: ExtraFormsandFormsetsMixin validate if `formset_list` or `extra_form_list` exists
+* Fix: App template tags, didn't got the model information intead of the app information
+
+
 0.1.0 (2016-06-12)
 ++++++++++++++++++
 * Fix: CRUD Messages Mixin